moved config model to separate file
